refactor(pattern): update admin queries & tests to match new patterns
