+++ src/admin/resolvers/queries/IABCategory.auth.integration.ts
@@ -1,25 +1,13 @@
-import { db, getServer } from '../';
+import { db, getServer } from '../../../test/admin-server';
 import {
   clear as clearDb,
   createIABCategoryHelper,
   getServerWithMockedHeaders,
-} from '../../helpers';
-import { GET_IAB_CATEGORIES } from './queries.gql';
-import {
-  ACCESS_DENIED_ERROR,
-  COLLECTION_CURATOR_FULL,
-  READONLY,
-} from '../../../shared/constants';
+} from '../../../test/helpers';
+import { GET_IAB_CATEGORIES } from './sample-queries.gql';
+import { ACCESS_DENIED_ERROR, READONLY } from '../../../shared/constants';
 
-describe('queries: IABCategory', () => {
-  const headers = {
-    name: 'Test User',
-    username: 'test.user@test.com',
-    groups: `group1,group2,${COLLECTION_CURATOR_FULL}`,
-  };
-
-  const server = getServerWithMockedHeaders(headers);
-
+describe('auth: IABCategory', () => {
   beforeAll(async () => {
     await clearDb(db);
   });
@@ -46,58 +34,6 @@ describe('queries: IABCategory', () => {
       );
       await createIABCategoryHelper(db, 'Coronavirus', iabParent3);
       await createIABCategoryHelper(db, 'Fitness', iabParent3);
-    });
-
-    it('should get IAB categories in alphabetical order', async () => {
-      const {
-        data: { getIABCategories: data },
-      } = await server.executeOperation({
-        query: GET_IAB_CATEGORIES,
-      });
-
-      // Even though we've created several IAB categories, we should only
-      // receive the three parent ones back
-      expect(data.length).toEqual(3);
-
-      // The parent categories should be in alphabetical order
-      expect(data[0].name).toEqual('Food and Drink');
-      expect(data[1].name).toEqual('Health and Wellness');
-      expect(data[2].name).toEqual('Technology');
-
-      // And so should the child categories of each parent IAB category
-      // "Food and Drink"
-      expect(data[0].children.length).toEqual(2);
-      expect(data[0].children[0].name).toEqual('Chocolate');
-      expect(data[0].children[1].name).toEqual('Pizza');
-
-      // "Health and Wellness"
-      expect(data[1].children.length).toEqual(2);
-      expect(data[1].children[0].name).toEqual('Coronavirus');
-      expect(data[1].children[1].name).toEqual('Fitness');
-
-      // "Technology"
-      expect(data[2].children.length).toEqual(3);
-      expect(data[2].children[0].name).toEqual('Internet');
-      expect(data[2].children[1].name).toEqual('Self-driving Cars');
-      expect(data[2].children[2].name).toEqual('Wearables');
-    });
-
-    it('should get all available properties of IAB categories', async () => {
-      const {
-        data: { getIABCategories: data },
-      } = await server.executeOperation({
-        query: GET_IAB_CATEGORIES,
-      });
-
-      // all the props of the first parent IAB category
-      expect(data[0].externalId).toBeTruthy();
-      expect(data[0].name).toBeTruthy();
-      expect(data[0].slug).toBeTruthy();
-
-      // and all the props of its first child category
-      expect(data[0].children[0].externalId).toBeTruthy();
-      expect(data[0].children[0].name).toBeTruthy();
-      expect(data[0].children[0].slug).toBeTruthy();
     });
 
     it('should succeed if a user has only READONLY access', async () => {
