+++ src/admin/resolvers/queries/Language.integration.ts
@@ -1,0 +1,30 @@
+import { getServerWithMockedHeaders } from '../../../test/helpers';
+import { GET_LANGUAGES } from './sample-queries.gql';
+import { COLLECTION_CURATOR_FULL } from '../../../shared/constants';
+import { CollectionLanguage } from '../../../database/types';
+
+describe('queries: Language', () => {
+  describe('getLanguages query', () => {
+    it('should get all available languages', async () => {
+      const headers = {
+        name: 'Test User',
+        username: 'test.user@test.com',
+        groups: `group1,group2,${COLLECTION_CURATOR_FULL}`,
+      };
+
+      const server = getServerWithMockedHeaders(headers);
+
+      const {
+        data: { getLanguages: data },
+      } = await server.executeOperation({
+        query: GET_LANGUAGES,
+      });
+
+      expect(data.length).toEqual(Object.values(CollectionLanguage).length);
+
+      data.forEach((language) => {
+        expect(language in CollectionLanguage).toBeTruthy();
+      });
+    });
+  });
+});
